Split sync runner tests
